refactor routes and db queries and successfully start server
