--- server/lib/user-queries.js
@@ -1,1 +1,0 @@
-const pool = require('./db');
